requirements.txt
streamlit
groq
pdfplumber
fpdf2
python-dotenv
